Travis: fix the build
The Random Compat repo introduced two files, one which is specifically for Psalm, another for PHPUnit, neither is compatible with PHP 5.2.

These files are `export-ignore`-d in their `.gitattributes`, so will never be part of a distribution, so we can safely ignore them when testing the ruleset against the latest version of the polyfill.