GH Actions: update the workflow to allow for Sodium Compat 2.x
Sodium Compat has tagged new releases in the 1.x series, but has now also released a 2.0 version, which has a PHP 8.1 minimum.

The GH Actions workflow for this package tests that the ruleset doesn't throw any errors on the polyfills repos themselves. So, to allow for both the 1.x as well as the 2.x branches, some changes to the workflows are required.

This commit sorts this out by:
* Adding extra builds to make sure that the SodiumCompat ruleset is tested against both Sodium Compat 1.x as well as 2.x.
* Making sure that the tests against the 2.x branch sets the PHPCompatibility `testVersion` to `8.1-`.

Refs:
* https://github.com/paragonie/sodium_compat/releases
* https://paragonie.com/blog/2024/04/release-sodium-compat-v2-and-future-our-polyfill-libraries