Ruleset: update for PHP 8.5
PHP 8.5 deprecated the `disabled_classes` ini setting, but the code in the `lib/random` file will never be executed on PHP >= 7.0, so this is not an issue.
As of the next version of PHPCompatibility, this will be flagged.

This commit updates the ruleset to disregard the warning when running PHPCompatibility over the code of the random compat package itself.